Move unit tests to its own file
