+++ snowpark-checkpoints-hypothesis/test/unit/__init__.py
@@ -1,0 +1,3 @@
+#
+# Copyright (c) 2012-2024 Snowflake Computing Inc. All rights reserved.
+#
